refactor university (Ilya groups): SSAU, PGUPS
