--- src/main/java/com/techstud/sch_parser/service/MappingService.java
@@ -1,17 +1,0 @@
-package com.techstud.sch_parser.service;
-
-import com.techstud.sch_parser.exception.EmptyScheduleException;
-import com.techstud.sch_parser.model.Schedule;
-import com.techstud.sch_parser.model.api.response.bmstu.BmstuApiResponse;
-import com.techstud.sch_parser.model.api.response.sseu.SseuApiResponse;
-import com.techstud.sch_parser.model.api.response.tltsu.TltsuApiResponse;
-import org.jsoup.nodes.Document;
-
-import java.util.List;
-
-public interface MappingService {
-
-    Schedule mapSsauToSchedule(List<Document> documents);
-
-    Schedule mapPgupsToSchedule(List<Document> documents);
-}
