+++ src/test/java/com/techstud/sch_parser/domain/SpbstuParserTest.java
@@ -4,9 +4,7 @@ package com.techstud.sch_parser.domain;
 import com.techstud.sch_parser.domain.impl.SpbstuParser;
 import com.techstud.sch_parser.model.Schedule;
 import com.techstud.sch_parser.model.kafka.request.ParsingTask;
-import com.techstud.sch_parser.service.MappingService;
 import com.techstud.sch_parser.service.MappingServiceRef;
-import com.techstud.sch_parser.service.impl.MappingServiceImpl;
 import com.techstud.sch_parser.service.impl.SpbstuServiceImpl;
 import lombok.extern.log4j.Log4j2;
 import org.jsoup.nodes.Document;
