+++ src/test/java/com/techstud/sch_parser/domain/TltsuParserTest.java
@@ -3,9 +3,7 @@ package com.techstud.sch_parser.domain;
 import com.techstud.sch_parser.domain.impl.TltsuParser;
 import com.techstud.sch_parser.model.api.response.tltsu.TltsuApiResponse;
 import com.techstud.sch_parser.model.kafka.request.ParsingTask;
-import com.techstud.sch_parser.service.MappingService;
 import com.techstud.sch_parser.service.MappingServiceRef;
-import com.techstud.sch_parser.service.impl.MappingServiceImpl;
 import com.techstud.sch_parser.service.impl.TItsuServiceImpl;
 import org.apache.http.impl.client.CloseableHttpClient;
 import org.apache.http.impl.client.DefaultHttpClient;
